+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added:
+- Integration tests using Fluttium
+
+### Changed:
+- Reordered fields in habit config
+- Default story removed from habits
+
+## [0.8.309] - 2023-04-08
 ### Changed:
 - Refactoring in audio recording